+++ LiteOrm.Demo/Services/DepartmentService.cs
@@ -1,7 +1,7 @@
-﻿using LiteOrm;
+using LiteOrm;
 using LiteOrm.Common;
 using LiteOrm.Demo.Models;
-using LiteOrm.Framework;
+using LiteOrm.DependencyInjection;
 using LiteOrm.Service;
 using Microsoft.Extensions.DependencyInjection;