+++ LiteOrm.Tests/Infrastructure/TestServices.cs
@@ -1,6 +1,6 @@
 using LiteOrm;
 using LiteOrm.Common;
-using LiteOrm.Framework;
+using LiteOrm.DependencyInjection;
 using LiteOrm.Service;
 using LiteOrm.Tests.Models;
 using Microsoft.Extensions.DependencyInjection;